Factor out SingleSmallNumericSearchContext from SingleSmallNumericAttribute.
