+++ searchlib/src/vespa/searchlib/attribute/CMakeLists.txt
@@ -110,6 +110,7 @@ vespa_add_library(searchlib_attribute OBJECT
     singlestringpostattribute.cpp
     single_numeric_enum_search_context.cpp
     single_numeric_search_context.cpp
+    single_small_numeric_search_context.cpp
     single_string_enum_search_context.cpp
     single_string_enum_hint_search_context.cpp
     sourceselector.cpp
